// ui/src/generated/Keymap.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { KeyBinding } from "./KeyBinding";
import type { KeymapPreset } from "./KeymapPreset";

export type Keymap = { preset: KeymapPreset, 
/**
 * Sorted by action, one entry per action.
 */
overrides: Array<KeyBinding>, };